# Backend/openapi.yaml
openapi: 3.1.0
info:
  title: PulseBoard API
  version: 1.0.0
  description: |
    The same API the PulseBoard web app uses.

    **Auth.** Sign up with `POST /auth/register`, confirm the emailed code with
    `POST /auth/verify-email`, then call `POST /auth/login` to get a short-lived access token and send it as
    `Authorization: Bearer <token>`. The refresh token is set as an httpOnly cookie
    scoped to `/api/v1/auth`; call `POST /auth/refresh` to rotate it and get a new
    access token.

    **Rate limits.** Auth endpoints allow 10 requests per minute per IP. Everything else
    allows 300 per minute per user, and AI endpoints 20 per minute per user. Responses
    carry `RateLimit-Limit` and `RateLimit-Remaining`; a 429 includes `Retry-After`.

    **Errors** always look like `{ "error": { "message": "...", "code": "..." } }`.

    **Realtime.** Connect with Socket.io to the same host, passing `auth: { token }`.
    Emit `board:join` with a board id to receive `task:created`, `task:updated`,
    `task:deleted`, `comment:created`, `presence` and `cursor:move` events for it.
servers:
  - url: /api/v1
...
security:
  - bearer: []
paths:
...
  /workspaces/{workspaceId}/boards:
    parameters: [{ $ref: '#/components/parameters/workspaceId' }]
    get:
      tags: [Boards]
      summary: List boards in a workspace
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  boards:
                    type: array
                    items: { $ref: '#/components/schemas/Board' }
    post:
      tags: [Boards]
      summary: Create a board
      description: Requires **admin**.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, key]
              properties:
                name: { type: string }
                key: { type: string, pattern: '^[A-Z][A-Z0-9]{1,5}$', example: PLAT }
                description: { type: string }
      responses:
        '201': { description: Created }
        '403': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }
  /boards/{boardId}:
    parameters: [{ $ref: '#/components/parameters/boardId' }]
    get:
      tags: [Boards]
      summary: Full board state
      description: The board, all of its tasks, the workspace's members and the workflow rules.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  board: { $ref: '#/components/schemas/Board' }
                  tasks:
                    type: array
                    items: { $ref: '#/components/schemas/Task' }
                  members:
                    type: array
                    items: { $ref: '#/components/schemas/Member' }
                  role: { type: string, enum: [owner, admin, member] }
                  workflow:
                    type: object
                    properties:
                      statuses:
                        type: array
                        items: { $ref: '#/components/schemas/Status' }
                      transitions:
                        type: object
                        additionalProperties:
                          type: array
                          items: { $ref: '#/components/schemas/Status' }
    patch:
      tags: [Boards]
      summary: Update a board
      description: Requires **admin**.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                description: { type: string }
      responses:
        '200': { description: OK }
    delete:
      tags: [Boards]
      summary: Delete a board and everything on it
      description: Requires **admin**.
      responses:
        '204': { description: Deleted }
  /boards/{boardId}/tasks:
    parameters: [{ $ref: '#/components/parameters/boardId' }]
    post:
      tags: [Tasks]
      summary: Create a task
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TaskInput' }
      responses:
        '201': { $ref: '#/components/responses/Task' }
  /tasks/{taskId}:
    parameters: [{ $ref: '#/components/parameters/taskId' }]
    get:
      tags: [Tasks]
      summary: Get a task
      responses:
        '200': { $ref: '#/components/responses/Task' }
    patch:
      tags: [Tasks]
      summary: Update task fields
      description: Status changes go through `POST /tasks/{taskId}/move`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - { $ref: '#/components/schemas/TaskInput' }
                - type: object
                  properties:
                    dependsOn:
                      type: array
                      items: { type: string }
      responses:
        '200': { $ref: '#/components/responses/Task' }
        '400': { $ref: '#/components/responses/Error' }
    delete:
      tags: [Tasks]
      summary: Delete a task
      description: Admins can delete any task; members only tasks they created.
      responses:
        '204': { description: Deleted }
        '403': { $ref: '#/components/responses/Error' }
  /tasks/{taskId}/move:
    parameters: [{ $ref: '#/components/parameters/taskId' }]
    post:
      tags: [Tasks]
      summary: Change status and/or position
      description: |
        Enforces the board workflow. A task needs an assignee to start, can't start or
        finish while it depends on open tasks, and can only reach `done` from `in_review`.
        Violations return `422 invalid_transition` with a readable message.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status, position]
              properties:
                status: { $ref: '#/components/schemas/Status' }
                position: { type: number }
      responses:
        '200': { $ref: '#/components/responses/Task' }
        '422': { $ref: '#/components/responses/Error' }
...
  /tasks/{taskId}/attachments/upload-url:
    parameters: [{ $ref: '#/components/parameters/taskId' }]
    post:
      tags: [Attachments]
      summary: Get a signed URL to upload a file directly to storage
      description: PUT the file to `uploadUrl` with the returned headers, then confirm it with `POST /tasks/{taskId}/attachments`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, contentType, size]
              properties:
                name: { type: string }
                contentType: { type: string }
                size: { type: integer, maximum: 26214400 }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  key: { type: string }
                  uploadUrl: { type: string }
                  headers: { type: object }
        '503': { $ref: '#/components/responses/Error' }
  /tasks/{taskId}/attachments:
    parameters: [{ $ref: '#/components/parameters/taskId' }]
    post:
      tags: [Attachments]
      summary: Confirm an upload and attach it to the task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [key, name]
              properties:
                key: { type: string }
                name: { type: string }
      responses:
        '201': { $ref: '#/components/responses/Task' }
  /tasks/{taskId}/attachments/{attachmentId}/url:
    parameters:
      - { $ref: '#/components/parameters/taskId' }
      - { name: attachmentId, in: path, required: true, schema: { type: string } }
    get:
      tags: [Attachments]
      summary: Get a short-lived download URL
      responses:
        '200': { description: OK }
...
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    workspaceId: { name: workspaceId, in: path, required: true, schema: { type: string } }
    boardId: { name: boardId, in: path, required: true, schema: { type: string } }
    taskId: { name: taskId, in: path, required: true, schema: { type: string } }
  responses:
...
    Task:
      description: OK
      content:
        application/json:
          schema:
            type: object
            properties:
              task: { $ref: '#/components/schemas/Task' }
    Error:
      description: Error
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            message: { type: string }
            code: { type: string }
    User:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        username: { type: string }
        email: { type: string }
        color: { type: string }
    Member:
      allOf:
        - { $ref: '#/components/schemas/User' }
        - type: object
          properties:
            role: { type: string, enum: [owner, admin, member] }
...
    Board:
      type: object
      properties:
        id: { type: string }
        workspaceId: { type: string }
        name: { type: string }
        key: { type: string }
        description: { type: string }
    Status:
      type: string
      enum: [backlog, todo, in_progress, blocked, in_review, done]
    Priority:
      type: string
      enum: [none, low, medium, high, urgent]
    TaskInput:
      type: object
      properties:
        title: { type: string }
        description: { type: string }
        status: { $ref: '#/components/schemas/Status' }
        priority: { $ref: '#/components/schemas/Priority' }
        assigneeId: { type: [string, 'null'] }
        labels:
          type: array
          items: { type: string }
        dueDate: { type: [string, 'null'], format: date-time }
    Task:
      type: object
      properties:
        id: { type: string }
        boardId: { type: string }
        key: { type: string, example: PLAT-42 }
        number: { type: integer }
        title: { type: string }
        description: { type: string }
        status: { $ref: '#/components/schemas/Status' }
        priority: { $ref: '#/components/schemas/Priority' }
        assigneeId: { type: [string, 'null'] }
        labels:
          type: array
          items: { type: string }
        dueDate: { type: [string, 'null'], format: date-time }
        dependsOn:
          type: array
          items: { type: string }
        position: { type: number }
        attachments:
          type: array
          items:
            type: object
            properties:
              id: { type: string }
              name: { type: string }
              size: { type: integer }
              contentType: { type: string }
              uploadedAt: { type: string, format: date-time }
        createdBy: { type: string }
        startedAt: { type: [string, 'null'], format: date-time }
        completedAt: { type: [string, 'null'], format: date-time }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }